Update Queen.java
Next, we created a constructor for Queen, and moved the instance variable to the inside of it. In doing so, it is observable that &this.name& is valid code, since it isn't underlined or suggesting that there may be some kind of error when running; which tells us that super.name might work, and that variable assignment outside of the constructor failed because it is only possible for it to be defined inside of the constructor, probably. So, once we try to access &name&, which is the variable from &class Piece&.,. from the constructor..., only in that environment will the access become possible; and so we can use the variable &name& from the Piece class now! Maybe we can call it directly!!. !! !!!! is where we are going with this. So next we will assign this.name to super.name; .. since that is our goal in this particular stage of the puzzle. We will make that step in our next commit.